implemented 3 file structure
