+++ crypt_app.cpp
@@ -1,0 +1,5 @@
+#include"crypt_app.h"
+orderbookentry:: orderbookentry(double price,double amount,std::string timestamp, std::string product,ordertype type):
+                 price{price},amount{amount},product{product},timestamp{timestamp},type{type}{};
+
+
